refactor: テーマに関連するファイルをresource.themeフォルダに集約
