+++ app/routes/resources.theme/services/system-theme.client.ts
